+++ internal/handlers/balance/mock_storage.go
@@ -35,11 +35,11 @@ func (m *MockAbstrStorage) EXPECT() *MockAbstrStorageMockRecorder {
 }
 
 // SelectBalanceByLogin mocks base method.
-func (m *MockAbstrStorage) SelectBalanceByLogin(arg0 context.Context, arg1 string) (int, int, error) {
+func (m *MockAbstrStorage) SelectBalanceByLogin(arg0 context.Context, arg1 string) (float64, float64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "SelectBalanceByLogin", arg0, arg1)
-	ret0, _ := ret[0].(int)
-	ret1, _ := ret[1].(int)
+	ret0, _ := ret[0].(float64)
+	ret1, _ := ret[1].(float64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
